refactor(Point): Point 패키지 위치를 view 패키지로 변경
