+++ src/main/java/view/OutputView.java
@@ -2,7 +2,6 @@ package view;
 
 import domain.ladder.Ladder;
 import domain.ladder.Line;
-import domain.ladder.Point;
 import domain.player.Player;
 
 import java.util.List;
