+++ src/main/java/view/Point.java
@@ -1,6 +1,4 @@
-package domain.ladder;
-
-import utils.ErrorMessage;
+package view;
 
 import java.util.Arrays;
 
@@ -8,6 +6,8 @@ public enum Point {
 
     TRUE(true, "-----"),
     FALSE(false, "     ");
+
+    private static final String POINT_NOT_FOUND_ERROR = "일치하는 포인트 객체가 존재하지 않습니다.";
 
     private final Boolean flag;
     private final String message;
@@ -21,7 +21,7 @@ public enum Point {
         return Arrays.stream(values())
                 .filter(point -> point.flag == flag)
                 .findAny()
-                .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.POINT_NOT_FOUND_ERROR.getMessage()));
+                .orElseThrow(() -> new IllegalArgumentException(POINT_NOT_FOUND_ERROR));
     }
 
     public String getMessage() {
